guru: encoding fixes in docs, basic symbol information.

+++ docs/clean_docs.py
@@ -1,15 +1,16 @@
 import os
 
-from lxml import html
+from lxml import html, etree
 from lxml.html.clean import Cleaner
 import io
 import re
 
 
 def clean(inp):
-    source = html.parse(inp)
+    with open(inp, 'rb') as f:
+        source = html.fromstring(f.read(), parser=html.HTMLParser(encoding='utf-8', default_doctype=False))
 
-    root = source.getroot()
+    root = source
     for (element, attribute, link, pos) in root.iterlinks():
         if '.' not in link or not attribute == 'href':
             continue
@@ -18,20 +19,16 @@ def clean(inp):
         if ext in ('xml', 'html', 'xhtml'):
             element.set(attribute, symbol)
 
-    cleaner = Cleaner(remove_tags=['body'], kill_tags=['head'], comments=False)
+    cleaner = Cleaner(
+        remove_tags=['body'],
+        kill_tags=['head'],
+        comments=True,
+        processing_instructions=True,
+        remove_unknown_tags=False
+    )
     cleaned = cleaner.clean_html(source)
 
-    data = io.BytesIO()
-    cleaned.write(data, encoding='utf-8', method='html')
-
-    cleaned = data.getvalue().decode('utf-8')
-    cleaned = strip_xml_tag(cleaned)
-
-    return cleaned
-
-
-def strip_xml_tag(data):
-    return re.sub(r'<\?xml[^>]+>', '', data)
+    return etree.tounicode(cleaned, method='html')
 
 
 def main():
@@ -41,15 +38,18 @@ def main():
     # parser.add_argument('--output', type=argparse.FileType('w'), default=sys.stdout)
     # parser.add_argument('input')
     parser.add_argument('dir')
+    parser.add_argument('--dry', action='store_true', default=False)
     ns = parser.parse_args()
 
     for name in os.listdir(ns.dir):
         path = os.path.join(ns.dir, name)
         data = clean(path)
 
-        os.unlink(path)
-        with open(os.path.splitext(path)[0] + '.html', 'wb') as fobj:
-            fobj.write(data.encode('utf-8'))
+        if not ns.dry:
+            os.unlink(path)
+
+        with open(os.path.splitext(path)[0] + '.html', 'w', encoding='utf-8') as fobj:
+            fobj.write(data)
 
 
 if __name__ == '__main__':
